fix typo and align method name, move replace function to companion object
